Refactor: 도메인 기반 구조로 코드 개편
