+++ src/domains/chat/constant.ts
