+++ src/domains/chat/types.d.ts
@@ -1,0 +1,15 @@
+import type { Chat } from '@prisma/client';
+
+type SocketMessageType = 'read' | 'write' | 'close';
+export interface SocketMessage {
+  message_type: SocketMessageType;
+}
+export interface ReadChatRequest extends SocketMessage { // 클라이언트 -> 서버
+}
+export interface WriteChatRequest extends SocketMessage { // 클라이언트 -> 서버
+  content: string;
+}
+export type MessageBody = Chat // 서버 -> 클라이언트
+
+export interface CloseChatRequest extends SocketMessage {
+}
