+++ src/domains/chat/validator.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const WriteChatRequestSchema = z.object({
+  content: z.string().max(255),
+});
