+++ src/domains/review/types.d.ts
@@ -1,0 +1,12 @@
+import { UserReview } from '@prisma/client';
+
+export interface UserReviewResponse {
+  score: number;
+  review_note: string;
+  merit: string[],
+  created_at: Date;
+}
+
+export type CreateUserReviewRequestBody =
+  Pick<UserReview, 'score' | 'review_note'>
+  & { merit: string[] };
