+++ src/index.ts
@@ -1,20 +1,20 @@
 import express, { NextFunction, Request, Response } from 'express';
 import 'dotenv/config';
-import { kakaoLoginRoute } from './routes/kakaoLoginRoute';
-import { userRoute } from '@/routes/user';
+import { kakaoLoginRoute } from '@/domains/auth/route';
+import { userRoute } from '@/domains/user/route';
 import cors from 'cors';
 import fs from 'node:fs';
 import https from 'https';
 import * as path from 'node:path';
 import { sendError } from '@/utils/response';
 import { StatusCodes } from 'http-status-codes';
-import { matchRoute } from '@/routes/match';
-import { alarmRoute } from './routes/alarmRoute';
-import { userReviewRoute } from './routes/userReviewRoute';
-import { bugReportRoute } from '@/routes/bugReport';
+import { matchRoute } from '@/domains/match/route';
+import { alarmRoute } from '@/domains/alarm/route';
+import { bugReportRoute } from '@/domains/bugReport/route';
 import { WebSocketServer } from 'ws';
-import { handleSocketConnection } from '@/socket';
 import * as http from 'node:http';
+import { route } from '@/domains/review/route';
+import { handleSocketConnection } from '@/domains/chat/socket';
 
 const app = express();
 const DEV_PORT = 3000;
@@ -39,7 +39,7 @@ app.use(API_PREFIX, bugReportRoute);
 app.use(API_PREFIX, userRoute);
 app.use(API_PREFIX, matchRoute);
 app.use(API_PREFIX, alarmRoute);
-app.use(API_PREFIX, userReviewRoute);
+app.use(API_PREFIX, route);
 
 app.get('*', (req: Request, res: Response) => {
   if (req.path.startsWith('/api')) {
