--- src/routes/userReviewRoute.ts
@@ -1,9 +1,0 @@
-import express from 'express';
-import { hunterReviewController } from '@/controllers/userReviewController';
-import { authenticateToken } from '@/middlewares/authMiddleware';
-
-const router = express.Router();
-
-router.post('/review/:matchId', authenticateToken, hunterReviewController);
-
-export const userReviewRoute = router;
