--- src/types/websocket.d.ts
@@ -1,3 +1,0 @@
-import { WebSocket } from 'ws';
-
-export type NetworkSessions = { [id: number]: WebSocket };
